added timeout to cgi

+++ src/HttpRequest_Handler_Get.cpp
@@ -103,9 +103,9 @@ void	HttpRequest::continueDownload()
 		if (debug)std::cout <<response<<std::endl;
 		
 		
-		size_t bytesSent = send(_client_fd, response.c_str(), response.length(), 0);
+		long bytesSent = send(_client_fd, response.c_str(), response.length(), 0);
 
-		if (bytesSent != response.length())
+		if (bytesSent != (long)response.length())
 		{
 			if (debug)std::cout << BG_BRIGHT_RED<<"Sending header error" << RESET<<std::endl;
 			_state._errorOcurred = SEND_ERROR;
@@ -116,14 +116,14 @@ void	HttpRequest::continueDownload()
 	if (init == 1)
 	{
 		_state._downloadFile.read(responseBuffer, MAX_SEND_BYTES);
-		size_t bytesRead = _state._downloadFile.gcount();
+		int bytesRead = _state._downloadFile.gcount();
 		if (_state._downloadFile.bad())
 		{
 			if (debug)std::cout << BG_BRIGHT_RED<<"File reading error" << RESET<<std::endl;
 			_state._errorOcurred = READ_ERROR;
 			return;
 		}
-		size_t bytesSent = send(_client_fd, responseBuffer, bytesRead, 0);
+		int bytesSent = send(_client_fd, responseBuffer, bytesRead, 0);
 		totalSent += bytesRead;
 		if (bytesSent != bytesRead)
 		{
